+++ CHANGELOG.md
@@ -16,6 +16,7 @@ New methods:
 New features and improvements:
 
 - Added ``exclusive`` parameter to ``sjoin_nearest`` method for Shapely >= 2.0 (#2877)
+- Added ``metadata`` parameter to ``GeoDataFrame.to_file`` (#2850)
 
 Bug fixes:
 - Fix ambiguous error when GeoDataFrame is initialised with a column called "crs" (#2944)